Moved Displaying Native Balance to its own component
